ref: move Color to EMS\Helpers\Standard\Color
